requirements.txt
streamlit==1.28.0
numpy==1.24.3
joblib==1.3.2
deepface==0.0.79
Pillow==10.0.1
scikit-learn==1.3.0
plotly==5.15.0
opencv-python==4.8.1.78
tensorflow==2.13.0
retina-face==0.0.13